PyTorch/requirements.txt
absl-py==1.4.0
asttokens==3.0.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
dm-tree==0.1.8
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
humanfriendly==10.0
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.8.1
keras==2.15.0
kiwisolver==1.4.8
libclang==18.1.1
markdown==3.8.2
markupsafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
ml-dtypes==0.2.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
onnx==1.18.0
onnx-ir==0.1.4
onnx2pytorch==0.5.1
onnxruntime==1.22.1
onnxscript==0.3.2
opt-einsum==3.4.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt-toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyzmq==27.0.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-model-optimization==0.8.0
termcolor==3.1.0
tf-keras==2.15.1
tf2onnx==1.8.4
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
traitlets==5.14.3
triton==3.3.1
typing-extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.14.1
